requirements.txt
pip==22.2.2
PyYAML==6.0
setuptools==65.0.2
wheel==0.37.1
XlsxWriter==3.0.3
